fin de la sección 13 server actions

+++ src/app/api/todos/route.ts
@@ -1,4 +1,5 @@
 import prisma from '@/app/lib/prisma'
+import { getUserSessionServer } from '@/auth/actions/auth-actions'
 import { NextResponse, NextRequest } from 'next/server'
 
 import * as yup from 'yup'
@@ -34,9 +35,13 @@ const postSchema = yup.object({
 })
 export async function POST(request: Request) { 
 
+    const user = await getUserSessionServer()
+
+    if(!user) return NextResponse.json({error: 'no autorizado'}, {status: 401})
+
     try {
         const {complete, description} = await postSchema.validate( await request.json())
-        const todo = await prisma.todo.create({data: {complete, description}})
+        const todo = await prisma.todo.create({data: {complete, description, userId: user.id}})
         return NextResponse.json(todo)
     } catch (error) {
         return NextResponse.json({error}, {status: 400})
@@ -48,9 +53,14 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) { 
 
+    const user = await getUserSessionServer()
+
+    if(!user) return NextResponse.json({error: 'no autorizado'}, {status: 401})
+
+
     try {
          
-         await prisma.todo.deleteMany({where: {complete:true}})
+         await prisma.todo.deleteMany({where: {complete:true, userId:user.id}})
         return NextResponse.json('message: todos deleted')
     } catch (error) {
         return NextResponse.json({error}, {status: 400})
